Refactored db util
