+++ backend/dbHandler.js
@@ -1,0 +1,13 @@
+const mongoose = require("mongoose");
+
+mongoose.Promise = global.Promise;
+
+module.exports.connect = async (uri) => {
+
+    const mongooseOpts = {
+        useNewUrlParser: true,
+        useUnifiedTopology: true
+    };
+
+    return await mongoose.connect(uri, mongooseOpts);
+}
