+++ backend/index.js
@@ -2,7 +2,10 @@ const express = require('express');
 const bodyParser = require("body-parser");
 const app = express();
 const PORT = process.env.PORT || 5000;
-const db = require("./models");
+
+const db = require('./dbHandler');
+
+dbUrl = process.env.DBURI || "mongodb://localhost:27017/test";
 
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
@@ -12,14 +15,7 @@ app.get('/', function(req, res) {
     res.send('Hello world');
 });
 
-
-//Connection to Database
-db.mongoose
-  .connect(db.url, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true
-  })
-  .then(() => {
+db.connect(dbUrl).then(() => {
     console.log("Connected to the database!");
   })
   .catch(err => {
