--- backend/models/index.js
@@ -1,8 +1,0 @@
-const mongoose = require("mongoose");
-mongoose.Promise = global.Promise;
-
-const db = {};
-db.mongoose = mongoose;
-db.url = "mongodb://localhost:27017/test";
-
-module.exports = db;
