+++ backend/tests/integration/utils/testDbHandler.js
@@ -2,20 +2,14 @@
 
 const mongoose = require('mongoose');
 const { MongoMemoryServer } = require('mongodb-memory-server');
+const db = require('../../../dbHandler');
 
 const mongod = new MongoMemoryServer();
 
 module.exports.connect = async () => {
     const uri = await mongod.getConnectionString();
 
-    const mongooseOpts = {
-        useNewUrlParser: true,
-        autoReconnect: true,
-        reconnectTries: Number.MAX_VALUE,
-        reconnectInterval: 1000
-    };
-
-    await mongoose.connect(uri, mongooseOpts);
+    await db.connect(uri);
 }
 
 module.exports.closeDatabase = async () => {
